mtd: nand/bbt: allow scanning bad block markers in raw mode
The bad block markers, present in the OOB area, are currently read in
normal mode (MTD_OPS_PLACE_OOB).
This approach works in most cases since the implementation ignores ECC
errors.
However, with the addition of data randomizers/scramblers (required by
some MLC NANDs), we might have to read the BBM in raw mode.

This is the case on Allwinner platforms: Allwinner does not use the same
randomizer seeds for all its partitions (the primary bootloader has to use
an hardcoded seed which is not appropriate for the whole NAND chip), which
means you might end up with unproperly de-randomized BBM leading to
incorrect bad block detection.
On sunxi, reading data in raw mode disables both the ECC engine and the
randomizer which solves the problem.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Signed-off-by: Richard Weinberger <richard@nod.at>