drm: Add a mode_valid to the encoder
The bridges are usually attached to the encoder, which means that they will
entirely bypass the connectors' mode_valid, even though the limitations
might actually be at the encoder level (for example because it's not able
to generate a pixel clock high enough).

In such a case, we want to filter out the modes exposed to the user by
filtering them using a connector-like mode_fixup.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>