drm/sun4i: DO NOT UPSTREAM: Deal with the broken VGA adapter
The VGA bridge doesn't care about the data enable signal, and the display
engine sends data during the sync periods. This is bad, because we'll end
up sending data during the VGA syncs, that are going to be used later on by
monitors to calibrate on black.

This essentially means that if we put a colored pixel at the end of any
line, its colors will be ignored for the next line.

We can however take advantage of the fact that the background when there's
no plane in some coordinate is black. Using this, we can just shift all the
planes by one pixel on the right, meaning that we will always have a black
pixel as the last pixel.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>